1.修复在Oracle,SQLServer下的一些问题. 2.添加部分数据库的单元测试

+++ mybatis-nosugar-api/src/main/java/com/nosugarice/mybatis/mapper/select/SelectPageMapper.java
@@ -140,7 +140,7 @@ public interface SelectPageMapper<T> extends SelectMapper {
      */
     default Page<T> selectPage(Page<T> page, FunS<List<T>> selectFunction, Object... params) {
         Assert.notNull(page, "Page不能为空.");
-        long count = page.getTotal() > 0 ? page.getTotal() : adapterCount(selectFunction, params);
+        long count = page.getTotal() > 0 ? page.getTotal() : adapterCount(selectFunction, params).longValue();
         page.setTotal(count);
         if (count > 0 && (long) (page.getNumber() - 1) * page.getSize() < count) {
             try {
